pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-velocity-dashboard"
version = "0.1.0"
description = "AI Velocity Dashboard - Monitor your AI development team's productivity, test coverage, and infrastructure health."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "AI Team", email = "ai-team@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "streamlit>=1.32.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.1",
    "fastapi>=0.109.2",
    "uvicorn>=0.27.0",
    "PyGithub>=2.1.1",
    "langsmith>=0.0.87",
    "pandas>=2.1.4",
    "numpy>=1.26.3",
    "plotly>=5.18.0",
    "matplotlib>=3.8.2",
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.13.2",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
    "types-python-dateutil>=2.8.19",
    "types-requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/your-org/ai-velocity-dashboard"
Documentation = "https://github.com/your-org/ai-velocity-dashboard#readme"
Issues = "https://github.com/your-org/ai-velocity-dashboard/issues"
Source = "https://github.com/your-org/ai-velocity-dashboard"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=app --cov-report=term-missing --cov-report=html"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["app"]
omit = ["**/__init__.py", "**/tests/**"]

[tool.coverage.report]
show_missing = true
skip_covered = true

[project.scripts]
ai-velocity-dashboard = "app.main:main"
